+++ packages/expo-font/ios/FontUtilsRecords.swift
@@ -5,6 +5,6 @@ import ExpoModulesCore
 
 struct RenderToImageOptions: Record {
   @Field var fontFamily: String = ""
-  @Field var size: Float = 24
+  @Field var size: CGFloat = 24
   @Field var color: Color = .black
 }
